Extract PackageReference into separate file
